revert to image replacements function
