+++ packages/exports/src/gSuite/docs/findPlaceholderIndex.ts
@@ -1,7 +1,7 @@
 import type { docs_v1 } from "@googleapis/docs";
 
 /**
- * Helper to find the index of a placeholder (including Markdown syntax) in the document.
+ * Helper to find the index of a placeholder in the document.
  */
 export async function findPlaceholderIndex(
   googleDocs: docs_v1.Docs,
@@ -14,42 +14,29 @@ export async function findPlaceholderIndex(
   if (!body) return null;
 
   for (const element of body) {
-    // Handle table elements
     if (element.table) {
       for (const row of element.table.tableRows || []) {
         for (const cell of row.tableCells || []) {
           for (const cellElement of cell.content || []) {
-            const index = findInParagraph(cellElement, placeholder);
-            if (index !== null) return index;
+            if (cellElement.paragraph?.elements) {
+              for (const textElement of cellElement.paragraph.elements) {
+                if (textElement.textRun?.content?.includes(placeholder)) {
+                  return textElement.startIndex ?? null;
+                }
+              }
+            }
           }
         }
       }
     }
 
-    // Handle regular paragraphs
-    if (element.paragraph) {
-      const index = findInParagraph(element, placeholder);
-      if (index !== null) return index;
+    if (element.paragraph?.elements) {
+      for (const textElement of element.paragraph.elements) {
+        if (textElement.textRun?.content?.includes(placeholder)) {
+          return textElement.startIndex ?? null;
+        }
+      }
     }
   }
-
-  return null;
-}
-
-/**
- * Helper to find placeholder in a paragraph or similar structure.
- */
-function findInParagraph(
-  element: docs_v1.Schema$StructuralElement,
-  placeholder: string,
-): number | null {
-  if (!element.paragraph?.elements) return null;
-
-  for (const textElement of element.paragraph.elements) {
-    if (textElement.textRun?.content?.includes(placeholder)) {
-      return textElement.startIndex ?? null;
-    }
-  }
-
   return null;
 }
